fix(sampler): centralize prompt history policy
