+++ crates/sipp/src/tests/runtime/config/inference_config_tests.rs
@@ -171,6 +171,77 @@ fn sampling_patch_merges_only_common_knobs() {
 }
 
 #[test]
+fn prompt_sampler_seed_start_uses_default_penalty_tail() {
+    let config = NativeRuntimeConfig::default();
+
+    assert_eq!(config.prompt_sampler_seed_start(None, 100), 36);
+    assert_eq!(config.prompt_sampler_seed_start(None, 32), 0);
+}
+
+#[test]
+fn prompt_sampler_seed_start_respects_sampling_overrides() {
+    let mut config = NativeRuntimeConfig::default();
+    config.sampling.repeat_last_n = Some(7);
+    let patch = RequestSampling::Patch(SamplingRuntimePatch {
+        temperature: Some(0.2),
+        top_p: None,
+    });
+    let full = RequestSampling::Full(SamplingRuntimeConfig {
+        repeat_last_n: Some(11),
+        ..SamplingRuntimeConfig::default()
+    });
+
+    assert_eq!(config.prompt_sampler_seed_start(Some(&patch), 20), 13);
+    assert_eq!(config.prompt_sampler_seed_start(Some(&full), 20), 9);
+}
+
+#[test]
+fn prompt_sampler_seed_start_preserves_full_history_modes() {
+    let repeat_all = RequestSampling::Full(SamplingRuntimeConfig {
+        repeat_last_n: Some(-1),
+        ..SamplingRuntimeConfig::default()
+    });
+    let dry_all = RequestSampling::Full(SamplingRuntimeConfig {
+        samplers: vec![SamplerStage::Dry],
+        dry_multiplier: Some(0.8),
+        dry_penalty_last_n: None,
+        ..SamplingRuntimeConfig::default()
+    });
+    let config = NativeRuntimeConfig::default();
+
+    assert_eq!(config.prompt_sampler_seed_start(Some(&repeat_all), 100), 0);
+    assert_eq!(config.prompt_sampler_seed_start(Some(&dry_all), 100), 0);
+}
+
+#[test]
+fn prompt_sampler_seed_start_skips_inactive_history_samplers() {
+    let no_history = RequestSampling::Full(SamplingRuntimeConfig {
+        samplers: vec![SamplerStage::TopK, SamplerStage::TopP],
+        ..SamplingRuntimeConfig::default()
+    });
+    let disabled_penalty = RequestSampling::Full(SamplingRuntimeConfig {
+        repeat_last_n: Some(0),
+        repeat_penalty: Some(1.0),
+        ..SamplingRuntimeConfig::default()
+    });
+    let mirostat = RequestSampling::Full(SamplingRuntimeConfig {
+        mirostat: Some(1),
+        ..SamplingRuntimeConfig::default()
+    });
+    let config = NativeRuntimeConfig::default();
+
+    assert_eq!(
+        config.prompt_sampler_seed_start(Some(&no_history), 100),
+        100
+    );
+    assert_eq!(
+        config.prompt_sampler_seed_start(Some(&disabled_penalty), 100),
+        100
+    );
+    assert_eq!(config.prompt_sampler_seed_start(Some(&mirostat), 100), 100);
+}
+
+#[test]
 fn native_runtime_config_normalizes_policy_limits() {
     let mut config = NativeRuntimeConfig::default();
     config.context.n_parallel = Some(0);
